backend/requirements.txt
flask
flask-cors
psutil
requests
python-dotenv